requirements.txt
discord.py
confight
matplotlib
pandas
pid
